+++ src/compas_ghpython/artists/lineartist.py
@@ -35,23 +35,6 @@ class LineArtist(PrimitiveArtist):
         return compas_ghpython.draw_lines(lines)[0]
 
     @staticmethod
-    def draw_collection(collection):
-        """Draw the collection of lines.
-
-        Parameters
-        ----------
-        collection : list of compas.geometry.Line
-            A collection of ``Line`` objects.
-
-        Returns
-        -------
-        list of :class:`Rhino.Geometry.Line`
-
-        """
-        lines = [LineArtist._get_args(primitive) for primitive in collection]
-        return compas_ghpython.draw_lines(lines)
-
-    @staticmethod
     def _get_args(primitive):
         start = list(primitive.start)
         end = list(primitive.end)